move index_master_location_list and master_location_detail views to unsorted for now
